Move b, u helpers to util module for now
Later on we can likely nuke many uses of these
