+++ paramiko/ber.py
@@ -16,9 +16,9 @@
 # along with Paramiko; if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA.
 from paramiko.common import max_byte, zero_byte, byte_ord, byte_chr
-from paramiko.py3compat import b
 
 import paramiko.util as util
+from paramiko.util import b
 
 
 class BERException(Exception):
