+++ paramiko/packet.py
@@ -38,8 +38,7 @@ from paramiko.common import (
     zero_byte,
     byte_ord,
 )
-from paramiko.py3compat import u
-from paramiko.util import asbytes
+from paramiko.util import asbytes, u
 from paramiko.ssh_exception import SSHException, ProxyCommandFailure
 from paramiko.message import Message
