+++ tests/test_auth.py
@@ -35,7 +35,7 @@ from paramiko import (
     AuthenticationException,
 )
 from paramiko import AUTH_FAILED, AUTH_PARTIALLY_SUCCESSFUL, AUTH_SUCCESSFUL
-from paramiko.py3compat import u
+from paramiko.util import u
 
 from .loop import LoopSocket
 from .util import _support, slow
